+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 **All WoW Versions:**  
 - 1080p improvements (BigWigs and Details)  
 
+**Retail:**  
+- Added basic edit mode specialization strings for all 40 specs  
+- Added new Cooldown Manager section in /lucky config  
+
 ## Version 4.11 [29-04-2026]
 
 **All WoW Versions:**