+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 **All WoW Versions:**  
 - 1080p improvements (BigWigs and Details)  
 - Added optional profile for the Baganator addon  
+- Updated frFR locales (Credits: Oniria)  
+- Updated graphics config section mouseover descriptions  
 
 **Retail:**  
 - Added basic edit mode specialization strings for all 40 specs (WIP)